tests/requirements.txt
pytest
requests
jsonschema
pytest-xdist
pyyaml
pytest-html
